Lesson 1: command line parsing
I am going to start the port with what I *think* should be something
extremely simple yet satisfying, which is to implement dcraw's -i option.
This switch simply tells dcraw to read the camera's make and model from
the header information and print it to standard out.

The first step is to get any sort of argc/argv parsing in go to work. It
turns out this is pretty simple using the "flag" package.

  http://golang.org/pkg/flag/

Interestingly, go programs don't seem to get the explict argc and argv
that we all know and love in C, or at least if they do, I wasn't able
to find them. Instead, the flag package knows how to magically find
command line args.

Note also the init() function which is automatically run as part of
intializing "package main". This is a convenient place to stick these
sorts of setup functions.

Now let's do some comparisons:
  8 var identify_only bool
  9 func init() {
 10         flag.BoolVar(&identify_only, "i", false, "Identify files without decoding them")
 11 }
 12
 13 func main() {
 14         flag.Parse()

versus

9077 int CLASS main (int argc, const char **argv)
9078 {
...
9080   int timestamp_only=0, thumbnail_only=0, identify_only=0;
...
9145   argv[argc] = "";
9146   for (arg=1; (((opm = argv[arg][0]) - 2) | 2) == '+'; ) {
9147     opt = argv[arg++][1];
9148     if ((cp = (char *) strchr (sp="nbrkStqmHACg", opt)))
9149       for (i=0; i < "114111111422"[cp-sp]-'0'; i++)
9150         if (!isdigit(argv[arg+i][0])) {
9151           fprintf (stderr,_("Non-numeric argument to \"-%c\"\n"), opt);
9152           return 1;
9153         }
9154     switch (opt) {
...
9187       case 'i':  identify_only     = 1;  break;

And later on, this bit of go:
 16         if len(flag.Arg(0)) == 0 {
 17                 fmt.Println("No files to process.")
 18         }

Corresponds to:
9213   if (arg == argc) {
9214     fprintf (stderr,_("No files to process.\n"));
9215     return 1;
9216   }

Not sure if what I did was idiomatic go or not, but it works for now.

All in all, this is much easier than wrangling with getopt.

  http://www.gnu.org/software/libc/manual/html_node/Example-of-Getopt.html

TODO:
  - I'm printing out error conditions on stdout, not stderr
  - I'm not exiting with a return value
  - my strings aren't localizable yet