Lesson 3: tests
Our port of dcraw is humming along and we're properly reading bytes
out of our test file, so that's great. But when you take a look at the
original dcraw.c file, you might say that "hairy" would be a charitable
description at best.

As we re-implement the original in go and build confidence that we're
doing the right thing, it's important that we don't ever regress in
functionality.

So we'll use go's built-in testing framework to write little unit tests
for ourselves and later, as we descend into darkness, we'll have a tiny
little thread leading us back out into sanity.

I fumbled around for a bit here, and decided that the easiest way to
use the testing framework was to reorganize the source files and take
advantage of go's built-in conventions.

So now, we can:

  $ cd src/dcraw
  $ go build
  $ go test
  PASS
  ok  	_/home/achiang/Projects/dcraw/src/dcraw	0.005s

Observe that globals from the main program are available to us in the
test program:

 10         ifp, _ = os.Open("../../test/IMG_1100.CR2")

I am sure we will definitely need to refactor the code and tests at some
point, but until we understand dcraw.c a little better, it's best to just
keep the port as obvious as possible for now.